Initial commit
This contains the basic files that will be present for the ambiance
tool. At the moment this actually does nothing. It loads the file,
queues some data in a buffer and does not play the source or creates a
listener.

It also does not necessarily frees everything (the source in
particular, should be missing). Well this is just a WIP.

Signed-off-by: Guillaume Clement <gclement@baobob.org>